Project setup activity and minor IDE driven refactoring.
