+++ AUTHORS
@@ -1,1 +1,1 @@
-Some Author <email_address>
+Olivier Seret (path-help@sanger.ac.uk)
