--- README_template.md
@@ -1,81 +1,0 @@
-# MYPROJECT
-A single sentence describing what the software is/what the software does.
-
-[![Build Status](https://travis-ci.org/sanger-pathogens/seroba.svg?branch=master)](https://travis-ci.org/sanger-pathogens/seroba)   
-[![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-brightgreen.svg)](https://github.com/sanger-pathogens/seroba/blob/master/LICENSE)   
-[![status](https://img.shields.io/badge/MGEN-10.1099%2Fmgen.0.000056-brightgreen.svg)](http://mgen.microbiologyresearch.org/content/journal/mgen/10.1099/mgen.0.000186)   
-[![status](https://img.shields.io/badge/Bioinformatics-10.1093-brightgreen.svg)](https://doi.org/10.1093/bioinformatics/btw022)  
-[![status](https://img.shields.io/badge/GenomeBiology-10.1186-brightgreen.svg)](https://genomebiology.biomedcentral.com/articles/10.1186/s13059-015-0849-0)   
-[![install with bioconda](https://img.shields.io/badge/install%20with-bioconda-brightgreen.svg)](http://bioconda.github.io/recipes/seroba/README.html)  
-[![Container ready](https://img.shields.io/badge/container-ready-brightgreen.svg)](https://quay.io/repository/biocontainers/seroba)  
-[![Docker Build Status](https://img.shields.io/docker/build/sangerpathogens/seroba.svg)](https://hub.docker.com/r/sangerpathogens/seroba)  
-[![Docker Pulls](https://img.shields.io/docker/pulls/sangerpathogens/seroba.svg)](https://hub.docker.com/r/sangerpathogens/seroba)  
-[![codecov](https://codecov.io/gh/sanger-pathogens/mlst_check/branch/master/graph/badge.svg)](https://codecov.io/gh/sanger-pathogens/mlst_check) 
-
-## Contents (edit as fit)
-  * [Introduction](#introduction)
-  * [Installation](#installation)
-    * [Required dependencies](#required-dependencies)
-    * [Optional dependencies](#optional-dependencies)
-    * [Linux specific instructions (Debian, Ubuntu, RedHat etc\.)](#linux-specific-instructions-debian-ubuntu-redhat-etc)
-    * [Mac OS](#mac-os)
-    * [Bioconda](#bioconda)
-    * [Homebrew/Linuxbrew](#homebrewlinuxbrew)
-    * [Docker](#docker)
-    * [Virtual Machine](#virtual-machine)
-    * [Galaxy](#galaxy)
-    * [From Source](#from-source)
-    * [Running the tests](#running-the-tests)
-  * [Usage](#usage)
-  * [License](#license)
-  * [Feedback/Issues](#feedbackissues)
-  * [Citation](#citation)
-  * [Further Information](#further-information)
-
-## Introduction
-Provide a more in-depth overview and description of the software. A single paragraph should be sufficient.
-
-## Installation
-<software name> has the following dependencies:
-
-### Required dependencies
-List the required dependencies along with the minimal version.
-
-### Optional dependencies
-List the optional dependencies along with the minimal version.
-
-There are a number of ways to install <software name> and details are provided below. If you encounter an issue when installing <software name> please contact your local system administrator. If you encounter a bug please log it [here](link_to_github_issues_page) or email us at path-help@sanger.ac.uk <or appropriate tool email list e.g. iva@sanger.ac.uk>.
-
-Include installation instructions for the following where appropriate:
-
-### Linux specific instructions (Debian, Ubuntu, RedHat etc.) 
-### Mac OS
-### Bioconda
-### Homebrew/Linuxbrew
-### Docker
-### Virtual Machine
-### Galaxy
-### From Source
-
-### Running the tests
-Instructions on how to run the tests and check that the software has installed correctly.
-
-## Usage
-Usage/output of running the --help option for the software. After this, include an example usage of the software, the bare minimum to get up and running (like a quick start guide) and describe the required input data and output files produced. This should be brief. For more information see the website and Jupyter tutorial (add links to these if relevant).
-
-## License
-<software name> is free software, licensed under [<license>](link_to_license_file_on_github).
-
-## Feedback/Issues
-Please report any issues to the [issues page](link_to_github_issues_page) or email path-help@sanger.ac.uk <or appropriate tool email list e.g. iva@sanger.ac.uk>.
-
-## Citation
-If you use this software please cite:
-<Insert citation (journal publication, bioarxiv, JOSS or github repo)>
-
-Also include any additional references that should be cited.
-
-## Further Information (optional)
-For more information on this software see:
-* [Software Web page](link_to_web_page)
-* [Jupyter notebook tutorial](https://github.com/sanger-pathogens/pathogen-informatics-training)
