+++ pyobo/InputTypes.py
@@ -1,4 +1,3 @@
-import os
 import argparse
 
 class InputTypes:
