+++ pyobo/tests/dummy_test.py
@@ -1,5 +1,5 @@
 import unittest
-import os
+
 
 #test_modules_dir = os.path.dirname(os.path.realpath(__file__))
 #data_dir = os.path.join(test_modules_dir, 'data','read')
