+++ setup.py
@@ -1,32 +1,30 @@
-import os
-import shutil
-import sys
 import glob
+import os
+
 from setuptools import setup, find_packages
+
 
 def read(fname):
     return open(os.path.join(os.path.dirname(__file__), fname)).read()
 
+
 version = 'x.y.z'
 if os.path.exists('VERSION'):
-  version = open('VERSION').read().strip()
+    version = open('VERSION').read().strip()
 
 setup(
-    name='MYPROJECT',
+    name='pyobo',
     version=version,
-    description='What Tim did next',
-	long_description=read('README.md'),
-    packages = find_packages(),
-    author='Tim Stickland',
+    description='A parser of Open Biomedical Ontologies (OBO) files',
+    long_description=read('README.md'),
+    packages=find_packages(),
+    author='Olivier Seret',
     author_email='path-help@sanger.ac.uk',
-    url='https://github.com/trstickland/MYPROJECT',
+    url='https://github.com/sanger-pathogens/pyobo',
     scripts=glob.glob('scripts/*'),
     test_suite='nose.collector',
     tests_require=['nose >= 1.3'],
-    install_requires=[
-         'biopython >= 1.68',
-         #'pyfastaq >= 3.12.0'
-       ],
+    install_requires=[],
     license='GPLv3',
     classifiers=[
         'Development Status :: 4 - Beta',
